+++ src/main/java/com/nhnacademy/store99/bookstore/book_author/repository/impl/BookAuthorRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.store99.bookstore.book_author.repository.imp;
+package com.nhnacademy.store99.bookstore.book_author.repository.impl;
 
 import com.nhnacademy.store99.bookstore.author.entity.QAuthor;
 import com.nhnacademy.store99.bookstore.book.entity.Book;
@@ -32,8 +32,8 @@ import org.springframework.stereotype.Component;
  * @author yrrho2
  */
 @Component
-public class BookAuthorRepositoryImp extends QuerydslRepositorySupport implements BookAuthorRepositoryCustom {
-    public BookAuthorRepositoryImp() {
+public class BookAuthorRepositoryImpl extends QuerydslRepositorySupport implements BookAuthorRepositoryCustom {
+    public BookAuthorRepositoryImpl() {
         super(BookAuthor.class);
     }
