--- src/test/java/com/nhnacademy/store99/bookstore/book_author/repository/imp/BookAuthorRepositoryImpTest.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.store99.bookstore.book_author.repository.imp;
-
-class BookAuthorRepositoryImpTest {
-
-}
