+++ src/test/java/com/nhnacademy/store99/bookstore/book_author/repository/impl/BookAuthorRepositoryImpTest.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.store99.bookstore.book_author.repository.impl;
+
+class BookAuthorRepositoryImpTest {
+
+}
